Refactor dashboard layout navigation
Closes #1157
