+++ layouts/dashboard.vue
@@ -5,101 +5,17 @@
 
     <!-- Main Content Area -->
     <main class="flex-1 flex flex-col min-w-0 min-h-0 h-screen lg:h-auto">
-      <!-- Main Content Header -->
-      <header class="bg-surface border-b border-border px-4 py-3 md:px-8 md:py-4 flex-shrink-0">
-        <div class="flex items-center justify-between gap-3">
-          <div class="flex items-center gap-2 sm:gap-4 min-w-0">
-            <div class="min-w-0">
-              <h1 class="text-lg sm:text-xl md:text-3xl font-bold text-text-primary leading-tight truncate">
-                {{ animatedDisplayTitle }}
-                <span v-if="isTypingTitle" class="title-caret" aria-hidden="true"></span>
-              </h1>
-              <p v-if="dynamicLastUpdateText || displaySubtitle" class="text-xs text-muted-foreground mt-0.5 truncate">{{ dynamicLastUpdateText || displaySubtitle }}</p>
-            </div>
-          </div>
-
-          <TransitionGroup
-            name="header-actions"
-            tag="div"
-            class="relative flex items-center gap-1.5 md:gap-2 flex-shrink-0"
-          >
-            <!-- Notification Bell — desktop only -->
-            <NotificationsNotificationBell key="notifications-bell" class="hidden lg:flex" />
-
-            <!-- Global Header Actions -->
-            <NuxtLink
-              key="upload-invoice"
-              to="/abastecimiento/compras-directas/crear"
-              class="flex items-center gap-1 md:gap-2 h-11 bg-primary text-primary-foreground px-2 md:px-4 rounded-xl font-medium hover:bg-primary/90 transition-all"
-              title="Cargar Factura IA"
-            >
-              <svg xmlns="http://www.w3.org/2000/svg" width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M11.017 2.814a1 1 0 0 1 1.966 0l1.051 5.558a2 2 0 0 0 1.594 1.594l5.558 1.051a1 1 0 0 1 0 1.966l-5.558 1.051a2 2 0 0 0-1.594 1.594l-1.051 5.558a1 1 0 0 1-1.966 0l-1.051-5.558a2 2 0 0 0-1.594-1.594l-5.558-1.051a1 1 0 0 1 0-1.966l5.558-1.051a2 2 0 0 0 1.594-1.594z"/><path d="M20 2v4"/><path d="M22 4h-4"/><circle cx="4" cy="20" r="2"/></svg>
-              <span class="hidden sm:inline">Cargar Factura IA</span>
-            </NuxtLink>
-
-            <button
-              key="pos-link"
-              type="button"
-              @click="navigateToPOS"
-              class="flex items-center gap-1 md:gap-2 h-11 bg-white border-2 border-surface-secondary text-text-primary px-2 md:px-4 rounded-lg text-sm font-medium hover:bg-surface-secondary transition-colors"
-              title="Venta POS"
-            >
-              <svg xmlns="http://www.w3.org/2000/svg" width="18" height="18" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M5 12h14"/><path d="M12 5v14"/></svg>
-              <span class="hidden sm:inline">Venta POS</span>
-            </button>
-
-            <!-- Portal Target for Custom Actions (pages can still inject extra actions) -->
-            <div key="portal-actions" id="dashboard-header-actions" class="flex items-center"></div>
-
-            <!-- Header Action Button (e.g., Print) -->
-            <button
-              v-if="dynamicHeaderAction"
-              key="dynamic-header-action"
-              @click="dynamicHeaderAction.handler"
-              class="h-11 px-4 bg-primary text-primary-foreground rounded-lg font-medium hover:bg-primary/90 transition-colors flex items-center gap-2"
-            >
-              <svg v-if="dynamicHeaderAction.icon" class="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M17 17h2a2 2 0 002-2v-4a2 2 0 00-2-2H5a2 2 0 00-2 2v4a2 2 0 002 2h2m2 4h6a2 2 0 002-2v-4a2 2 0 00-2-2H9a2 2 0 00-2 2v4a2 2 0 002 2zm8-12V5a2 2 0 00-2-2H9a2 2 0 00-2 2v4h10z" />
-              </svg>
-              {{ dynamicHeaderAction.label }}
-            </button>
-
-            <!-- Status Badge -->
-            <span
-              v-if="dynamicStatus"
-              key="dynamic-status"
-              :class="['h-11 flex items-center px-3 rounded-lg text-sm font-medium', dynamicStatus.color]"
-            >
-              {{ dynamicStatus.label }}
-            </span>
-
-            <!-- Restaurant open/close toggle -->
-            <DashboardBusinessStatusToggle key="business-status-toggle" />
-
-            <!-- Refresh Button (Desktop only) - matrix on manual refresh and progressive/first load -->
-            <button
-              key="refresh-button"
-              @click="handleRefresh"
-              :disabled="isRefreshing || isProgressiveLoading"
-              aria-label="Refrescar datos"
-              :aria-busy="isRefreshing || isProgressiveLoading"
-              class="hidden md:flex w-11 h-11 items-center justify-center bg-surface-secondary border-0 rounded-lg text-primary transition-all focus:outline-none focus:ring-2 focus:ring-ring disabled:opacity-50 disabled:cursor-not-allowed"
-              title="Refrescar">
-              <UiLoadingMatrix v-if="isRefreshing || isProgressiveLoading" size="5.5px" />
-              <svg v-else
-                class="w-5 h-5 transition-transform duration-300 hover:rotate-180"
-                fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
-                  d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15">
-                </path>
-              </svg>
-            </button>
-
-            <!-- Tenant Selector (Desktop only) -->
-            <DashboardTenantSelector key="tenant-selector" />
-          </TransitionGroup>
-        </div>
-      </header>
+      <DashboardShellHeader
+        :title="animatedDisplayTitle"
+        :subtitle="dynamicLastUpdateText || displaySubtitle"
+        :is-typing-title="isTypingTitle"
+        :status="dynamicStatus"
+        :header-action="dynamicHeaderAction"
+        :is-refreshing="isRefreshing"
+        :is-progressive-loading="isProgressiveLoading"
+        @refresh="handleRefresh"
+        @navigate-pos="navigateToPOS"
+      />
 
       <!-- Subscription Banner -->
       <SubscriptionBanner
@@ -136,23 +52,14 @@
       </div>
     </main>
 
-  <!-- Bottom Navigation - Mobile Only -->
-  <div
-    class="lg:hidden fixed bottom-0 inset-x-0 z-50 bg-white border-t border-titan-300 shadow-lg"
-    style="padding-bottom: env(safe-area-inset-bottom, 0px)"
-  >
-    <PosCartBottomBar
-      v-if="showPosMobileCartBar"
-      :visible="posMobileCartItemCount > 0"
-      :item-count="posMobileCartItemCount"
-      :formatted-total="posMobileCartFormattedTotal"
-      @open-cart="openPosMobileCart"
-    />
-    <DashboardBottomNav
+    <DashboardMobileChrome
       :active-page="activePage"
       :notifications-count="notificationsUnreadCount"
+      :show-pos-cart-bar="showPosMobileCartBar"
+      :pos-cart-item-count="posMobileCartItemCount"
+      :pos-cart-formatted-total="posMobileCartFormattedTotal"
+      @open-cart="openPosMobileCart"
     />
-  </div>
 
     <!-- Mobile Order Toast — client-only to avoid SSR/Teleport hydration mismatch -->
     <ClientOnly>
@@ -165,14 +72,13 @@
 </template>
 
 <script setup lang="ts">
-import { provide, inject, ref, computed, watch, onMounted, onUnmounted, type Ref, type ComputedRef } from 'vue'
+import { provide, ref, computed, watch, onMounted, onUnmounted } from 'vue'
 import {
   ChevronRightIcon
 } from '@heroicons/vue/24/outline'
 import { useNotifications } from '~/composables/useNotifications'
 import { useBilling } from '~/composables/useBilling'
 import { usePosMobileCart } from '~/composables/usePosMobileCart'
-import { WAREHOUSE_COPY } from '~/constants/warehouseCopy'
 
 // Notifications — init here so SSE starts on all screen sizes (not just when bell mounts)
 const { unreadCount: notificationsUnreadCount, init: initNotifications, disconnect: disconnectNotifications } = useNotifications()
@@ -196,860 +102,14 @@ const goBack = () => {
   }
 }
 
-// Determine configuration based on route
-const getPageConfig = () => {
-  const path = route.path
-
-  if (path === '/financiero' || path === '/dashboard') {
-    return {
-      pageTitle: 'Dashboard',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar métricas financieras...',
-      activePage: 'financiero' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/financiero/tir') {
-    return {
-      pageTitle: 'Análisis TIR',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar métricas TIR...',
-      activePage: 'financiero' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/financiero/analisis') {
-    return {
-      pageTitle: 'Análisis de Productos',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar productos...',
-      activePage: 'financiero' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/financiero/obstaculos') {
-    return {
-      pageTitle: 'Obstáculos del TIR',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar obstáculos...',
-      activePage: 'financiero' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/abastecimiento' || path === '/abastecimiento/') {
-    return {
-      pageTitle: 'Dashboard - Abastecimiento',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar en abastecimiento...',
-      activePage: 'abastecimiento' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/abastecimiento/proveedores') {
-    return {
-      pageTitle: 'Gestión de Proveedores',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar proveedores...',
-      activePage: 'abastecimiento' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/abastecimiento/calidad-datos') {
-    return {
-      pageTitle: 'Calidad de Datos',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'abastecimiento' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/abastecimiento/compras-directas' || path === '/abastecimiento/compras-directas/') {
-    return {
-      pageTitle: 'Compras Directas',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar compras...',
-      activePage: 'abastecimiento' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/abastecimiento/compras-directas/crear') {
-    return {
-      pageTitle: 'Nueva Compra Directa',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'abastecimiento' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: { label: 'Volver' }
-    }
-  } else if (path.startsWith('/abastecimiento/compras-directas/')) {
-    return {
-      pageTitle: 'Compra Directa',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'abastecimiento' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: { label: 'Volver' }
-    }
-  } else if (path === '/abastecimiento/ajustes' || path.startsWith('/abastecimiento/ajustes/')) {
-    return {
-      pageTitle: path.includes('/crear') ? 'Ajustar stock' : 'Historial de ajustes',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'abastecimiento' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: { label: 'Volver' }
-    }
-  } else if (path === '/abastecimiento/stock') {
-    return {
-      pageTitle: 'Stock',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar stock...',
-      activePage: 'abastecimiento' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/abastecimiento/movimientos') {
-    return {
-      pageTitle: 'Movimientos',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar movimientos...',
-      activePage: 'abastecimiento' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/abastecimiento/ingredientes-propios') {
-    return {
-      pageTitle: WAREHOUSE_COPY.warehouseCatalog,
-      pageSubtitle: undefined,
-      searchPlaceholder: WAREHOUSE_COPY.catalogSearchPlaceholder,
-      activePage: 'abastecimiento' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/abastecimiento/compras') {
-    return {
-      pageTitle: 'Órdenes de Compra',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar órdenes...',
-      activePage: 'abastecimiento' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path.includes('/abastecimiento/compra/') && path.includes('/acciones')) {
-    return {
-      pageTitle: 'Acciones de la Orden',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'abastecimiento' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: {
-        label: 'Volver'
-      }
-    }
-  } else if (path.includes('/abastecimiento/compra/') && path.includes('/transicion')) {
-    // Transition pages open in new window, no back button needed
-    return {
-      pageTitle: 'Historial de Cambios',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'abastecimiento' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/abastecimiento/compra/crear') {
-    return {
-      pageTitle: 'Crear Nueva Orden de Compra',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'abastecimiento' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: {
-        label: 'Volver'
-      }
-    }
-  } else if (path === '/abastecimiento/proveedor/crear') {
-    return {
-      pageTitle: 'Crear Proveedor',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'abastecimiento' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: {
-        label: 'Volver'
-      }
-    }
-  } else if (path.startsWith('/abastecimiento/proveedor/') && path !== '/abastecimiento/proveedor/crear') {
-    return {
-      pageTitle: 'Editar Proveedor',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'abastecimiento' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: {
-        label: 'Volver'
-      }
-    }
-  } else if (path.startsWith('/abastecimiento/compra/')) {
-    return {
-      pageTitle: 'Editar Orden de Compra',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'abastecimiento' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: {
-        label: 'Volver'
-      }
-    }
-  } else if (path === '/pagos/registrar') {
-    return {
-      pageTitle: 'Registrar Pago',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'pagos' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: {
-        label: 'Volver'
-      }
-    }
-  } else if (path === '/pagos' || path === '/pagos/') {
-    return {
-      pageTitle: 'Gestión de Pagos',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar pagos...',
-      activePage: 'pagos' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path.startsWith('/operaciones/promociones')) {
-    return {
-      pageTitle: 'Promociones',
-      pageSubtitle: 'Descuentos y reglas para el menú',
-      searchPlaceholder: undefined,
-      activePage: 'operaciones' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path.startsWith('/operaciones')) {
-    return {
-      pageTitle: 'Operaciones',
-      pageSubtitle: 'Comandas, cocina y mesas',
-      searchPlaceholder: undefined,
-      activePage: 'operaciones' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path.startsWith('/equipo/miembros/') && path !== '/equipo/miembros' && path !== '/equipo/miembros/') {
-    return {
-      pageTitle: 'Perfil de mesero',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'equipo' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path.startsWith('/equipo')) {
-    return {
-      pageTitle: 'Equipo',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'equipo' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/integraciones') {
-    return {
-      pageTitle: 'Integraciones Dashboard',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'integraciones' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path.includes('/analitica/rentabilidad')) {
-    return {
-      pageTitle: 'Rentabilidad',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar métricas...',
-      activePage: 'analytics' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path.includes('/analitica/clientes')) {
-    return {
-      pageTitle: 'Clientes',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar clientes...',
-      activePage: 'analytics' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path.includes('/analitica/puntos')) {
-    return {
-      pageTitle: 'Puntos Waros',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'analytics' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path.includes('/analytics') || path.includes('/analitica')) {
-    return {
-      pageTitle: 'Mis Ventas',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar métricas...',
-      activePage: 'analytics' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/finanzas' || path === '/finanzas/') {
-    return {
-      pageTitle: 'Finanzas',
-      pageSubtitle: 'Gestión financiera del negocio',
-      searchPlaceholder: undefined,
-      activePage: 'finanzas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/finanzas/arqueo') {
-    return {
-      pageTitle: 'Arqueo de Caja',
-      pageSubtitle: 'Cierre diario de efectivo',
-      searchPlaceholder: undefined,
-      activePage: 'finanzas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/finanzas/arqueo/nuevo') {
-    return {
-      pageTitle: 'Nuevo Arqueo',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'finanzas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: { label: 'Volver' }
-    }
-  } else if (path === '/finanzas/arqueo/x') {
-    return {
-      pageTitle: 'Corte X',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'finanzas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: { label: 'Volver' }
-    }
-  } else if (path === '/finanzas/arqueo/z') {
-    return {
-      pageTitle: 'Arqueo por turno u horario',
-      pageSubtitle: 'Ventana personalizada de fechas y horas',
-      searchPlaceholder: undefined,
-      activePage: 'finanzas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: { label: 'Volver' }
-    }
-  } else if (path.startsWith('/finanzas/arqueo/')) {
-    return {
-      pageTitle: 'Detalle Arqueo',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'finanzas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: { label: 'Volver' }
-    }
-  } else if (path === '/finanzas/gastos') {
-    return {
-      pageTitle: 'Gastos',
-      pageSubtitle: 'Control de egresos',
-      searchPlaceholder: undefined,
-      activePage: 'finanzas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/finanzas/gastos/crear') {
-    return {
-      pageTitle: 'Registrar Gasto',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'finanzas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: { label: 'Volver' }
-    }
-  } else if (path.startsWith('/finanzas/gastos/')) {
-    return {
-      pageTitle: 'Detalle Gasto',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'finanzas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: { label: 'Volver' }
-    }
-  } else if (path === '/finanzas/cartera') {
-    return {
-      pageTitle: 'Cartera',
-      pageSubtitle: 'Cuentas por cobrar',
-      searchPlaceholder: undefined,
-      activePage: 'finanzas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/finanzas/metodos-pago') {
-    return {
-      pageTitle: 'Métodos de Pago',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'finanzas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path.startsWith('/finanzas/metodos-pago/')) {
-    return {
-      pageTitle: 'Grupo de Pago',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'finanzas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: { label: 'Volver' }
-    }
-  } else if (path === '/finanzas/contabilidad/cuentas') {
-    return {
-      pageTitle: 'Plan de Cuentas',
-      pageSubtitle: 'Contabilidad general',
-      searchPlaceholder: undefined,
-      activePage: 'finanzas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path.startsWith('/finanzas/contabilidad/cuentas/')) {
-    return {
-      pageTitle: 'Detalle Cuenta',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'finanzas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: { label: 'Volver' }
-    }
-  } else if (path === '/finanzas/contabilidad/asientos') {
-    return {
-      pageTitle: 'Asientos Contables',
-      pageSubtitle: 'Libro diario',
-      searchPlaceholder: undefined,
-      activePage: 'finanzas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/finanzas/contabilidad/asientos/crear') {
-    return {
-      pageTitle: 'Nuevo Asiento',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'finanzas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: { label: 'Volver' }
-    }
-  } else if (path === '/finanzas/contabilidad/balance-comprobacion') {
-    return {
-      pageTitle: 'Balance de Comprobación',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'finanzas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/finanzas/cierre-contable') {
-    return {
-      pageTitle: 'Cierre Contable',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'finanzas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/finanzas/reportes/pl-mensual') {
-    return {
-      pageTitle: 'P&L Mensual',
-      pageSubtitle: 'Estado de resultados',
-      searchPlaceholder: undefined,
-      activePage: 'finanzas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path.startsWith('/finanzas')) {
-    return {
-      pageTitle: 'Finanzas',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'finanzas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path.includes('/reportes')) {
-    return {
-      pageTitle: 'Dashboard - Reportes',
-      pageSubtitle: 'Generación y gestión de reportes',
-      searchPlaceholder: 'Buscar reportes...',
-      activePage: 'reportes' as const,
-      showBreadcrumb: true,
-      breadcrumbPage: 'Reportes',
-      backButton: undefined
-    }
-  } else if (path.includes('/configuracion')) {
-    return {
-      pageTitle: 'Dashboard - Configuración',
-      pageSubtitle: 'Configuración del sistema',
-      searchPlaceholder: 'Buscar configuración...',
-      activePage: 'configuracion' as const,
-      showBreadcrumb: true,
-      breadcrumbPage: 'Configuración',
-      backButton: undefined
-    }
-  } else if (path.includes('/gestion/billing')) {
-    return {
-      pageTitle: 'Mi Plan',
-      pageSubtitle: 'Gestión de tu suscripción y facturación',
-      searchPlaceholder: 'Buscar en billing...',
-      activePage: 'admin' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path.includes('/admin')) {
-    return {
-      pageTitle: 'Admin',
-      pageSubtitle: 'Panel de administración del sistema',
-      searchPlaceholder: 'Buscar en administración...',
-      activePage: 'admin' as const,
-      showBreadcrumb: true,
-      breadcrumbPage: 'Administración',
-      backButton: undefined
-    }
-  } else if (path.includes('/pos/producto/')) {
-    return {
-      pageTitle: 'Personalizar Producto',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'pos' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: {
-        label: 'Volver'
-      }
-    }
-  } else if (path === '/pos/checkout') {
-    return {
-      pageTitle: 'Confirmar Orden',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'pos' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: {
-        label: 'Volver'
-      }
-    }
-  } else if (path === '/pos' || path.includes('/pos')) {
-    return {
-      pageTitle: 'Punto de Venta',
-      pageSubtitle: undefined,
-      searchPlaceholder: undefined,
-      activePage: 'pos' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: {
-        label: 'Volver'
-      }
-    }
-  } else if (path === '/ventas' || path === '/ventas/') {
-    return {
-      pageTitle: 'Ventas',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar ventas...',
-      activePage: 'ventas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/menu/productos') {
-    return {
-      pageTitle: 'Productos',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar productos...',
-      activePage: 'menu' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/menu/recetas') {
-    return {
-      pageTitle: 'Recetas',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar recetas...',
-      activePage: 'menu' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/menu/modificadores') {
-    return {
-      pageTitle: 'Modificadores',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar modificadores...',
-      activePage: 'menu' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/menu' || path === '/menu/') {
-    return {
-      pageTitle: 'Menú',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar...',
-      activePage: 'menu' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/inventario/stock') {
-    return {
-      pageTitle: 'Stock de Inventario',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar ingredientes...',
-      activePage: 'inventario' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/inventario/movimientos') {
-    return {
-      pageTitle: 'Movimientos de Inventario',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar movimientos...',
-      activePage: 'inventario' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path.startsWith('/inventario/ajustes/crear')) {
-    return {
-      pageTitle: 'Ajustar Inventario',
-      pageSubtitle: 'Registra ajustes manuales al inventario',
-      searchPlaceholder: undefined,
-      activePage: 'inventario' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/inventario/ajustes') {
-    return {
-      pageTitle: 'Ajustes de Inventario',
-      pageSubtitle: 'Resumen e historial de ajustes manuales',
-      searchPlaceholder: 'Buscar ajustes...',
-      activePage: 'inventario' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/inventario' || path === '/inventario/') {
-    return {
-      pageTitle: 'Inventario',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar...',
-      activePage: 'inventario' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path.startsWith('/despacho')) {
-    return {
-      pageTitle: 'Despacho',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar pedidos...',
-      activePage: 'despacho' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/negocio') {
-    return {
-      pageTitle: 'Mi Negocio',
-      pageSubtitle: 'Información y configuración de tu negocio',
-      searchPlaceholder: 'Buscar en negocio...',
-      activePage: 'negocio' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/operaciones/bitacora' || path === '/operaciones/bitacora/') {
-    return {
-      pageTitle: 'Bitácora de operaciones',
-      pageSubtitle: 'Auditoría de acciones en POS — mesas, barra y mostrador',
-      searchPlaceholder: undefined,
-      activePage: 'operaciones' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/operaciones/turnos' || path === '/operaciones/turnos/') {
-    return {
-      pageTitle: 'Turnos',
-      pageSubtitle: 'Horarios reutilizables para arqueos de caja',
-      searchPlaceholder: undefined,
-      activePage: 'operaciones' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/operaciones/propinas' || path === '/operaciones/propinas/') {
-    return {
-      pageTitle: 'Configuración de propinas',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar...',
-      activePage: 'operaciones' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path === '/ventas/propinas' || path === '/ventas/propinas/') {
-    return {
-      pageTitle: 'Propinas',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar propinas...',
-      activePage: 'ventas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path.startsWith('/ventas')) {
-    return {
-      pageTitle: 'Ventas',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar ventas...',
-      activePage: 'ventas' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path.startsWith('/menu')) {
-    return {
-      pageTitle: 'Menú',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar en menú...',
-      activePage: 'menu' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path.startsWith('/integraciones')) {
-    return {
-      pageTitle: 'Integraciones',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar integraciones...',
-      activePage: 'integraciones' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  } else if (path.startsWith('/abastecimiento')) {
-    return {
-      pageTitle: 'Abastecimiento',
-      pageSubtitle: undefined,
-      searchPlaceholder: 'Buscar en abastecimiento...',
-      activePage: 'abastecimiento' as const,
-      showBreadcrumb: false,
-      breadcrumbPage: undefined,
-      backButton: undefined
-    }
-  }
-
-  return {
-    pageTitle: 'Dashboard',
-    pageSubtitle: undefined,
-    searchPlaceholder: 'Buscar...',
-    activePage: 'dashboard' as const,
-    showBreadcrumb: false,
-    breadcrumbPage: undefined,
-    backButton: undefined
-  }
-}
-
-const config = computed(() => getPageConfig())
-const pageTitle = computed(() => config.value.pageTitle)
-const pageSubtitle = computed(() => config.value.pageSubtitle)
-const searchPlaceholder = computed(() => config.value.searchPlaceholder)
-const activePage = computed(() => config.value.activePage)
-const showBreadcrumb = computed(() => config.value.showBreadcrumb)
-const breadcrumbPage = computed(() => config.value.breadcrumbPage)
-const backButton = computed(() => config.value.backButton)
-
-// Date and time functionality
-const currentDateTime = ref('')
-
-const updateDateTime = () => {
-  const now = new Date()
-  const options: Intl.DateTimeFormatOptions = {
-    weekday: 'long',
-    year: 'numeric',
-    month: 'long',
-    day: 'numeric',
-    hour: '2-digit',
-    minute: '2-digit',
-    timeZone: 'America/Bogota'
-  }
-  currentDateTime.value = now.toLocaleDateString('es-CO', options)
-}
-
-// Update time immediately and then every minute
-let dateTimeInterval: ReturnType<typeof setInterval> | null = null
+const {
+  pageTitle,
+  pageSubtitle,
+  activePage,
+  showBreadcrumb,
+  breadcrumbPage,
+  backButton,
+} = useDashboardPageConfig()
 
 // Redirect to billing portal when subscription is expired
 watch(accessStatus, (status) => {
@@ -1059,20 +119,16 @@ watch(accessStatus, (status) => {
 }, { immediate: true })
 
 onMounted(() => {
-  updateDateTime()
-  dateTimeInterval = setInterval(updateDateTime, 60000)
   if (process.client) initNotifications()
   fetchAccessStatus()
 })
 
 onUnmounted(() => {
-  if (dateTimeInterval) clearInterval(dateTimeInterval)
   disconnectNotifications()
 })
 
 // Refresh handler - shared via composable (provide/inject unreliable in Nuxt 3 layout↔page)
 const {
-  refreshHandler,
   isRefreshing,
   isProgressiveLoading,
   lastUpdateText: composableLastUpdateText,
@@ -1214,53 +270,6 @@ onUnmounted(() => {
 })
 </script>
 
-<style>
-.title-caret {
-  display: inline-block;
-  width: 0.08em;
-  height: 0.9em;
-  margin-left: 0.08em;
-  vertical-align: -0.08em;
-  background-color: currentColor;
-  animation: title-caret-blink 1s steps(1) infinite;
-}
-
-@keyframes title-caret-blink {
-  0%, 50% {
-    opacity: 1;
-  }
-
-  50.01%, 100% {
-    opacity: 0;
-  }
-}
-
-/* Hide empty portal target to prevent extra flex gaps */
-#dashboard-header-actions {
-  position: relative;
-}
-
-#dashboard-header-actions:empty {
-  display: none;
-}
-
-#dashboard-header-actions > * {
-  animation: dashboard-header-portal-in 0.28s cubic-bezier(0.22, 1, 0.36, 1);
-}
-
-@keyframes dashboard-header-portal-in {
-  from {
-    opacity: 0;
-    transform: translateX(14px);
-  }
-
-  to {
-    opacity: 1;
-    transform: translateX(0);
-  }
-}
-</style>
-
 <style scoped>
 /* Smooth sidebar navigation animation */
 .sidebar-nav {
@@ -1284,32 +293,6 @@ onUnmounted(() => {
   animation: contentFadeIn 0.5s ease-out;
 }
 
-/* Header action transitions */
-.header-actions-move,
-.header-actions-enter-active,
-.header-actions-leave-active {
-  transition:
-    transform 0.28s cubic-bezier(0.22, 1, 0.36, 1),
-    opacity 0.24s ease,
-    filter 0.24s ease;
-}
-
-.header-actions-enter-from {
-  opacity: 0;
-  filter: blur(2px);
-  transform: translateX(14px);
-}
-
-.header-actions-leave-to {
-  opacity: 0;
-  filter: blur(2px);
-  transform: translateX(-10px);
-}
-
-.header-actions-leave-active {
-  position: absolute;
-  pointer-events: none;
-}
 </style>
 
 <style scoped>
